Extract thrusters from App.tsx and create setPwm module.
