--- src/components/thrusters/SetPwmModule.js
@@ -1,11 +1,0 @@
-import SetPwm from './SetPwm';
-
-const SetPwmModule = (props) => {
-  return (
-    <div>
-      <SetPwm /> 
-    </div>
-  );
-};
-
-export default SetPwmModule;
