--- src/components/thrusters/ThustersModule.js
@@ -1,11 +1,0 @@
-import Controls from './Controls';
-
-const ThrustersModule = (props) => {
-  return (
-    <div>
-      <Controls /> 
-    </div>
-  );
-};
-
-export default ThrustersModule;
