+++ src/context/rosContext.tsx
@@ -2,5 +2,5 @@ import React from 'react';
 import * as ROSLIB from 'roslib';
 
 export const RosContext = React.createContext(new ROSLIB.Ros({
-    url: 'ws://192.168.0.31:9090'
+    url: 'ws://localhost:9090'
 }))
